* main.c (main): Cast arguments to print_status to be consitant 	with format string.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13363 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/clients/ksu/main.c
@@ -738,7 +738,7 @@ main (argc, argv)
 		     target_user, (long) target_pwd->pw_uid); 
     }else{
 	print_status("Changing uid to %s (%ld)\n", 
-		     target_user, target_pwd->pw_uid); 
+		     target_user, (long) target_pwd->pw_uid); 
     }
     
 #ifdef	HAVE_SETLUID
